+++ src/freelunch/sat/satLifter/Stopwatch.java
@@ -55,7 +55,7 @@ public class Stopwatch {
     }
     
     public float elapsedSecondsFloat() {
-        return ((float)elapsedNanoSeconds()) / 1e9f;
+        return (elapsedNanoSeconds()) / 1e9f;
     }
 
     public int elapsedSeconds() {
